refactored templates snippets
